% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_frequency.R
\name{build_frequency_table_generic_wide}
\alias{build_frequency_table_generic_wide}
\title{Convert long-format frequency table to wide format}
\usage{
build_frequency_table_generic_wide(freq_long)
}
\arguments{
\item{freq_long}{A long-format table with columns:
\code{habitat}, \code{distance}, \code{sampling}, \code{occurrence}.}
}
\value{
A wide-format data.frame with columns:
\code{distance},
\code{sampling_<habitat>},
\code{occurrence_<habitat>}.
}
\description{
Convert long-format frequency table to wide format
}
