% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_frequency.R
\name{compute_distance_frequency}
\alias{compute_distance_frequency}
\title{Compute frequency of distances within defined bins}
\usage{
compute_distance_frequency(dist_vec, breaks)
}
\arguments{
\item{dist_vec}{Numeric vector of distances.}

\item{breaks}{Numeric vector of class breakpoints.}
}
\value{
A data.frame with columns:
\code{distance} (bin midpoint),
\code{frequency} (percentage within each bin).
}
\description{
Compute frequency of distances within defined bins
}
