% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_distance.R
\name{compute_distance_to_feature}
\alias{compute_distance_to_feature}
\title{Compute distance from points to one habitat feature}
\usage{
compute_distance_to_feature(pts_sf, habitat_sf)
}
\arguments{
\item{pts_sf}{sf POINT.}

\item{habitat_sf}{sf object (POINT / LINESTRING / POLYGON, etc.).}
}
\value{
Numeric vector of minimum distances for each point.
}
\description{
Compute distance from points to one habitat feature
}
