% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_decay_models.R
\name{exponential_decay}
\alias{exponential_decay}
\title{Exponential decay model}
\usage{
exponential_decay(x, N0, b)
}
\arguments{
\item{x}{Numeric vector of distances.}

\item{N0}{Intercept parameter (frequency at distance 0).}

\item{b}{Decay rate parameter.}
}
\value{
Numeric vector of predicted frequency values.
}
\description{
Exponential decay model
}
